requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
alembic==1.18.4
pydantic==2.12.5
httpx==0.28.1
python-multipart==0.0.24
authlib==1.6.9
PyJWT[crypto]==2.12.1
dnspython==2.7.0
